refactor(ui): 画面 composable に状態を引数で受け取る版を追加
VRT が「実画面」を撮れるようにするための extract-parameter。振る舞いは変えていない。

- AppListScreen(appInfoList, onAppClick) / CalendarScreen(uiState) を追加し、
  ViewModel を受け取る既存版はそこへ委譲する
- CalendarScreen の縦センタリング Box は ViewModel 版にしか無く CalendarGrid には
  無いため、内側の composable を撮ると検証対象から漏れる。テスト側で AppTheme /
  Surface を再宣言する案は、画面の見た目の定義が二重化して main と乖離するため採らない
- private だった createPreviewAppList / createSampleCells を PreviewData.kt へ
  internal で移設し、@Preview と VRT で同じサンプルデータを共有する
